GUI: split into multiple files
makes changing things in the future a lot easier.

Fixes https://github.com/NIT-dgp/UIP/issues/261
Fixes https://github.com/NIT-dgp/UIP/issues/254
